Rename lorenz.py to lorenz-sz40.py
According to the BREAKING TELEPRINTER CIPHERS AT BLETCHLEY PARK: An edition of I. J. Good, D. Michie and G. Timms' GENERAL REPORT ON TUNNY WITH EMPHASIS ON STATISTICAL METHODS (1945) Edited and with introductions and notes by James A. Reeds, Whitfield Diffie and J. V. Field. This cipher is the Lorenz Schlüsselzusatz 40.